+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* 2020/09/09 - Bug Fix - Fixed issue where Actor's Owned Item images were not being exported correctly.
 * 2020/09/08 - Bug Fix - 0.7.x fix for Roll Table items reference updates.
 * 2020/09/05 - Bug Fix - Issue occured when updating references on an compendium actor's owned items if they don't have any.
 * 2020/09/04 - Bug Fix #90 - Export/Import now handles wildcard image randomizer.